Refactor imports and reorganize ProjectConfig class into project_config.py
